Move volume rules to ruleset
